+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ New features
 * Add :code:`modeling.map_eta_parameters` for creation of mappings between connected individual parameters, etas or omegas.
 * Add :code:`modeling.infer_datatypes` to infer and check if data columns can be converted a simpler datatype (e.g. int32)
 * Add :code:`modeling.set_n_transit_compartments`
+* Add :code:`modeling.create_basic_pd_model`
 * Add first version of a VPC tool
 
 Changes